requirements.txt
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-google-genai==0.0.11
langchain-groq==0.0.1
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0